% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_oom.R
\name{format_oom}
\alias{format_oom}
\title{Round and format number by order of magnitude}
\usage{
format_oom(x, oom)
}
\arguments{
\item{x}{A numeric.}

\item{oom}{A numeric.}
}
\value{
A character
}
\description{
This rounds numbers and fills up places with zeros according to the given order of magnitude.
}
\seealso{
Other rounding functions: \code{\link{oom}},
  \code{\link{round_up}}
}
