% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpformat.R
\name{mpformat}
\alias{mpformat}
\title{Round and format number by order of magnitude}
\usage{
mpformat(x, digits = oom(x))
}
\arguments{
\item{x}{A numeric.}

\item{digits}{A numeric.}
}
\value{
A character
}
\description{
This rounds numbers and fills up places with zeros according to the given order of magnitude.
}
\seealso{
Other rounding functions: \code{\link{oom}},
  \code{\link{round_up}}
}
