% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom.R
\name{oom}
\alias{oom}
\title{Get order of magnitude for numbers}
\usage{
oom(x)
}
\arguments{
\item{x}{A numeric vector of measurement uncertainties, e.g., sd, se}
}
\value{
A numeric vector containing orders of magnitude, can be passed to \code{\link[base:round]{base::round()}} as digits.
}
\description{
This is meant to get the order of magnitude from measurement errors to round these errors and measured values or centers like mean or median in an empirical way.
}
\details{
The following convention is implemented:
\itemize{
\item Round \code{x} to 2 significant digits.
\item If first significant digit is 1, return number of digits according to 2 significant digits.
\item Otherwise return number of digits according to 1 significant digit.
}
}
\examples{
# Round very long error to 1 significant digit
x <- 0.074385229
oom_x <- oom(x)
round(x, oom_x)

# This starts with 1 and gets 2 signif. digits
x <- 0.174385229
oom_x <- oom(x)
round(x, oom_x)

# This again, but rounds up to 1 signif. digit
x <- 0.197385229
oom_x <- oom(x)
round(x, oom_x)
}
\seealso{
Other rounding functions: \code{\link{mpformat}},
  \code{\link{round_up}}
}
