% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A numeric
}
\description{
Standard error of the mean
}
\examples{
# Get the SEM of some T (not t) distributed samples with ascending size
sem(rnorm(10, 50, 10))
sem(rnorm(20, 50, 10))
sem(rnorm(30, 50, 10))
}
